Split WASM heartbeat modules
